Move email settings to top of content section in editor
