+++ src/features/emails/components/EmailEditor/index.tsx
@@ -1,14 +1,16 @@
 import dynamic from 'next/dynamic';
-import { Alert, Box, useTheme } from '@mui/material';
+import { Alert, Box, FormControl, TextField, useTheme } from '@mui/material';
 import EditorJS, { OutputBlockData, OutputData } from '@editorjs/editorjs';
 import { FC, useEffect, useRef, useState } from 'react';
 
 import editorjsBlocksToZetkinBlocks from 'features/emails/utils/editorjsBlocksToZetkinBlocks';
 import EmailSettings from './EmailSettings';
 import messageIds from 'features/emails/l10n/messageIds';
-import { Msg } from 'core/i18n';
+import { Msg, useMessages } from 'core/i18n';
 import zetkinBlocksToEditorjsBlocks from 'features/emails/utils/zetkinBlocksToEditorjsBlocks';
 import { ZetkinEmail } from 'utils/types/zetkin';
+import useEmailSettings from 'features/emails/hooks/useEmailSettings';
+import useDebounce from 'utils/hooks/useDebounce';
 
 const EmailEditorFrontend = dynamic(() => import('./EmailEditorFrontend'), {
   ssr: false,
@@ -21,8 +23,12 @@ interface EmailEditorProps {
 
 const EmailEditor: FC<EmailEditorProps> = ({ email, onSave }) => {
   const theme = useTheme();
+  const messages = useMessages(messageIds);
   const apiRef = useRef<EditorJS | null>(null);
   const [selectedBlockIndex, setSelectedBlockIndex] = useState(0);
+  const { subject, emailAddress, setSubject, orgTitle } = useEmailSettings(
+    email.subject || ''
+  );
 
   const zetkinInitialContent = email.content
     ? JSON.parse(email.content)
@@ -58,6 +64,10 @@ const EmailEditor: FC<EmailEditorProps> = ({ email, onSave }) => {
     blocksRef.current = content.blocks;
   }, [content.blocks.length]);
 
+  const debouncedFinishedTyping = useDebounce(async (value: string) => {
+    onSave({ subject: value });
+  }, 400);
+
   return (
     <Box display="flex" flexDirection="column" height="100%">
       {readOnly && (
@@ -67,6 +77,38 @@ const EmailEditor: FC<EmailEditorProps> = ({ email, onSave }) => {
       )}
       <Box display="flex" height="100%">
         <Box flex={1} sx={{ overflowY: 'auto' }}>
+          <Box display="flex" gap={2} padding={2}>
+            <FormControl fullWidth sx={{ flex: 2 }}>
+              <TextField
+                disabled
+                fullWidth
+                label={messages.editor.settings.tabs.settings.senderNameInputLabel()}
+                size="small"
+                value={orgTitle}
+              />
+            </FormControl>
+            <FormControl fullWidth sx={{ flex: 2 }}>
+              <TextField
+                disabled
+                fullWidth
+                label={messages.editor.settings.tabs.settings.senderAddressInputLabel()}
+                size="small"
+                value={emailAddress}
+              />
+            </FormControl>
+            <TextField
+              disabled={readOnly}
+              fullWidth
+              label={messages.editor.settings.tabs.settings.subjectInputLabel()}
+              onChange={(event) => {
+                setSubject(event.target.value);
+                debouncedFinishedTyping(event.target.value);
+              }}
+              size="small"
+              sx={{ flex: 3 }}
+              value={subject}
+            />
+          </Box>
           <EmailEditorFrontend
             apiRef={apiRef}
             initialContent={{ blocks: initialContent }}
@@ -95,10 +137,8 @@ const EmailEditor: FC<EmailEditorProps> = ({ email, onSave }) => {
           <EmailSettings
             apiRef={apiRef}
             blocks={content.blocks}
-            onSave={onSave}
             readOnly={readOnly}
             selectedBlockIndex={selectedBlockIndex}
-            subject={email.subject || ''}
           />
         </Box>
       </Box>
